+++ package.json
@@ -37,5 +37,10 @@
     "library",
     "data",
     "operational"
-  ]
+  ],
+  "jest": {
+    "setupFiles": [
+      "<rootDir>/test/helpers.js"
+    ]
+  }
 }
