+++ test/snippet.test.js
@@ -1,17 +1,4 @@
 const { JSDOM } = require('jsdom')
-const fs = require('fs')
-const Mustache = require('mustache')
-
-function generateSnippet(functionName, token, accountName, dataSourceName, host) {
-  const snippet = fs.readFileSync('./src/snippet.js', { encoding: 'utf-8' })
-  return Mustache.render(snippet, {
-    functionName,
-    token,
-    accountName,
-    dataSourceName,
-    host
-  })
-}
 
 function generateDefaultDOM(functionName, token, accountName, dataSourceName, host) {
   return `
